Move login, signup, and logout routes outside of the protected part of the api
